+++ src/crawlee/crawlers/_playwright/_playwright_pre_nav_crawling_context.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from crawlee._types import BasicCrawlingContext
+from crawlee._types import BasicCrawlingContext, PageSnapshot
 from crawlee._utils.docs import docs_group
 
 if TYPE_CHECKING:
@@ -25,3 +25,20 @@ class PlaywrightPreNavCrawlingContext(BasicCrawlingContext):
 
     block_requests: BlockRequestsFunction
     """Blocks network requests matching specified URL patterns."""
+
+    async def get_snapshot(self) -> PageSnapshot:
+        """Get snapshot of crawled page."""
+        html = None
+        screenshot = None
+
+        try:
+            html = await self.page.content()
+        except Exception:
+            self.log.exception(f'Failed to get html snapshot for {self.request.url}.')
+
+        try:
+            screenshot = await self.page.screenshot(full_page=True, type='jpeg')
+        except Exception:
+            self.log.exception(f'Failed to get page screenshot for {self.request.url}.')
+
+        return PageSnapshot(html=html, screenshot=screenshot)