+++ src/crawlee/statistics/__init__.py
@@ -1,6 +1,6 @@
 # ruff: noqa: A005
 
-from ._models import FinalStatistics, StatisticsPersistedState, StatisticsState
+from ._models import FinalStatistics, StatisticsState
 from ._statistics import Statistics
 
-__all__ = ['FinalStatistics', 'Statistics', 'StatisticsPersistedState', 'StatisticsState']
+__all__ = ['FinalStatistics', 'Statistics', 'StatisticsState']