+++ src/crawlee/storage_clients/models.py
@@ -224,7 +224,7 @@ class UnprocessedRequest(BaseModel):
 
     model_config = ConfigDict(populate_by_name=True)
 
-    unique_key: Annotated[str, Field(alias='requestUniqueKey')]
+    unique_key: Annotated[str, Field(alias='uniqueKey')]
     url: Annotated[str, BeforeValidator(validate_http_url), Field()]
     method: Annotated[HttpMethod | None, Field()] = None