+++ tests/e2e/project_template/test_static_crawlers_templates.py
@@ -2,6 +2,7 @@ import os
 import re
 import subprocess
 from pathlib import Path
+from typing import Literal
 
 import pytest
 from apify_client import ApifyClientAsync
@@ -9,17 +10,42 @@ from cookiecutter.main import cookiecutter
 
 from crawlee._cli import default_start_url, template_directory
 from crawlee._utils.crypto import crypto_random_object_id
-from tests.e2e.project_template.utils import patch_crawlee_version_in_pyproject_toml_based_project
+from tests.e2e.project_template.utils import patch_crawlee_version_in_project
 
 # To run these tests locally, make sure you have apify-cli installed and available in the path.
 # https://docs.apify.com/cli/docs/installation
 
 
-@pytest.mark.parametrize('http_client', ['httpx', 'curl-impersonate'])
-@pytest.mark.parametrize('crawler_type', ['parsel', 'beautifulsoup'])
-@pytest.mark.parametrize('package_manager', ['uv', 'poetry'])
+@pytest.mark.parametrize(
+    'crawler_type',
+    [
+        pytest.param('playwright-camoufox', marks=pytest.mark.playwright_camoufox),
+        pytest.param('playwright', marks=pytest.mark.playwright),
+        pytest.param('parsel', marks=pytest.mark.parsel),
+        pytest.param('beautifulsoup', marks=pytest.mark.beautifulsoup),
+    ],
+)
+@pytest.mark.parametrize(
+    'http_client',
+    [
+        pytest.param('httpx', marks=pytest.mark.httpx),
+        pytest.param('curl-impersonate', marks=pytest.mark.curl_impersonate),
+    ],
+)
+@pytest.mark.parametrize(
+    'package_manager',
+    [
+        pytest.param('pip', marks=pytest.mark.pip),
+        pytest.param('uv', marks=pytest.mark.uv),
+        pytest.param('poetry', marks=pytest.mark.poetry),
+    ],
+)
 async def test_static_crawler_actor_at_apify(
-    tmp_path: Path, crawlee_wheel_path: Path, package_manager: str, crawler_type: str, http_client: str
+    tmp_path: Path,
+    crawlee_wheel_path: Path,
+    package_manager: Literal['pip', 'uv', 'poetry'],
+    crawler_type: str,
+    http_client: str,
 ) -> None:
     # Generate new actor name
     actor_name = f'crawlee-python-template-e2e-test-{crypto_random_object_id(8).lower()}'
@@ -40,8 +66,8 @@ async def test_static_crawler_actor_at_apify(
         output_dir=tmp_path,
     )
 
-    patch_crawlee_version_in_pyproject_toml_based_project(
-        project_path=tmp_path / actor_name, wheel_path=crawlee_wheel_path
+    patch_crawlee_version_in_project(
+        project_path=tmp_path / actor_name, wheel_path=crawlee_wheel_path, package_manager=package_manager
     )
 
     # Build actor using sequence of cli commands as the user would
@@ -56,8 +82,11 @@ async def test_static_crawler_actor_at_apify(
     build_process = subprocess.run(['apify', 'push'], capture_output=True, check=False, cwd=tmp_path / actor_name)  # noqa: ASYNC221, S603, S607
     # Get actor ID from build log
     actor_id_regexp = re.compile(r'https:\/\/console\.apify\.com\/actors\/(.*)#\/builds\/\d*\.\d*\.\d*')
-    # Why is it in stderr and not in stdout???
-    actor_id = re.findall(actor_id_regexp, build_process.stderr.decode())[0]
+
+    if match := re.findall(actor_id_regexp, build_process.stderr.decode()):
+        actor_id = match[0]
+    else:
+        raise AssertionError(f'Failed to find actor id in build log: {build_process.stderr.decode()}')
 
     client = ApifyClientAsync(token=os.getenv('APIFY_TEST_USER_API_TOKEN'))
     actor = client.actor(actor_id)
@@ -65,7 +94,7 @@ async def test_static_crawler_actor_at_apify(
     # Run actor
     try:
         assert build_process.returncode == 0
-        started_run_data = await actor.start()
+        started_run_data = await actor.start(memory_mbytes=8192)
         actor_run = client.run(started_run_data['id'])
 
         finished_run_data = await actor_run.wait_for_finish()
@@ -80,6 +109,6 @@ async def test_static_crawler_actor_at_apify(
     assert finished_run_data
     assert finished_run_data['status'] == 'SUCCEEDED', additional_run_info
     assert (
-        'Crawler.stop() was called with following reason: The crawler has reached its limit of 50 requests per crawl.'
+        'Crawler.stop() was called with following reason: The crawler has reached its limit of 10 requests per crawl.'
     ) in actor_run_log, additional_run_info
-    assert int(re.findall(r'requests_finished\s*│\s*(\d*)', actor_run_log)[-1]) >= 50, additional_run_info
+    assert int(re.findall(r'requests_finished\s*│\s*(\d*)', actor_run_log)[-1]) >= 10, additional_run_info