+++ tests/unit/sessions/test_models.py
@@ -5,7 +5,7 @@ from datetime import datetime, timedelta, timezone
 import pytest
 
 from crawlee.sessions._cookies import CookieParam
-from crawlee.sessions._models import SessionModel, SessionPoolModel
+from crawlee.sessions._models import SessionModel
 
 SESSION_CREATED_AT = datetime.now(timezone.utc)
 
@@ -78,33 +78,3 @@ def test_session_model(
 
     # Check that max_age is correctly parsed into a timedelta object
     assert session_direct.max_age == session_camel.max_age == session_snake.max_age == timedelta(minutes=30)
-
-
-def test_create_session_pool_with_direct_sessions(session_direct: SessionModel) -> None:
-    """Test creating a SessionPoolModel with direct session model instances."""
-    session_pool = SessionPoolModel(
-        persistence_enabled=False,
-        persist_state_kvs_name='test_session_pool',
-        persist_state_key='crawlee_session_pool_state',
-        max_pool_size=3,
-        session_count=0,
-        usable_session_count=0,
-        retired_session_count=0,
-        sessions=[session_direct],
-    )
-    assert session_pool.sessions == [session_direct]
-
-
-def test_create_session_pool_with_args_sessions(session_args_camel: dict, session_args_snake: dict) -> None:
-    """Test creating a SessionPoolModel using sessions initialized from camelCase and snake_case dicts."""
-    session_pool_camel = SessionPoolModel(
-        persistence_enabled=False,
-        persist_state_kvs_name='test_session_pool',
-        persist_state_key='crawlee_session_pool_state',
-        max_pool_size=3,
-        session_count=0,
-        usable_session_count=0,
-        retired_session_count=0,
-        sessions=[session_args_camel, session_args_snake],
-    )
-    assert session_pool_camel.sessions == [SessionModel(**session_args_camel), SessionModel(**session_args_snake)]